refatorando ponteiros

+++ SolutionShelves/src/Engine/Renderer/Buffer.h
@@ -33,7 +33,7 @@ namespace SolutionShelves
 	{
 		std::string Name;
 		ShaderDataType Type;
-		uint32_t Offset;
+		uint64_t Offset;
 		uint32_t Size;
 		bool Normalized;
